+++ src/block_writer.hpp
@@ -35,7 +35,7 @@
  * @return True if writing was successful, false otherwise (e.g., file error).
  */
 bool write_blocks_to_stream(const std::string& filename, uint32_t width,
-                            uint32_t height, uint16_t offset_bits,
+                            uint32_t height, uint32_t offset_bits,
                             uint16_t length_bits, bool adaptive, bool model,
                             const std::vector<Block>& blocks, bool binary_only);
